remix-tutorial: Add explicit build/index.js to start script
Previously in the `remix-tutorial` template the `start` script was set to `remix-serve build`. This is different than the standard "fresh install" Remix
start script, and when attempting to run it, results in a "Directory import ... is not supported resolving ES modules imported from <your project>/node_modules/@remix-run/serve/dist/cli.js"

Changing the template to utilize explicitly list the `./build/index.js` fixes this error when running the `start` command.